server/requirements.txt
fastapi
uvicorn[standard]
websockets
faster-whisper
numpy
python-dotenv
